+++ mill-crossplatform/src-mill0.11/com/github/lolgab/mill/crossplatform/VersionSpecific.scala
@@ -2,6 +2,8 @@ package com.github.lolgab.mill.crossplatform
 
 import mill.define.Cross
 import mill.scalalib._
+import mill.scalajslib._
+import mill.scalanativelib._
 
 private[crossplatform] object VersionSpecific {
   def getModules[M <: CrossPlatform](cross: Cross[M]) = cross.crossModules
@@ -12,11 +14,13 @@ private[crossplatform] object VersionSpecific {
       cross
     )
   }
+  trait CrossPlatformScalaModule extends ScalaModule {
+    private[crossplatform] protected def myArtifactNameParts: Seq[String]
+    override def artifactNameParts = myArtifactNameParts
+  }
   trait CrossPlatformCrossScalaModule extends CrossScalaModule {
     override def crossValue: String = myCrossValue
     private[crossplatform] protected def myCrossValue: String
-    override def artifactNameParts =
-      super.artifactNameParts().patch(crossWrapperSegments.size - 1, Nil, 1)
   }
   trait CrossPlatform extends Cross.Module[String] {
     override def crossValue: String = throw new Exception(
@@ -25,5 +29,18 @@ private[crossplatform] object VersionSpecific {
     )
 
     private[crossplatform] protected def myCrossValue: String = crossValue
+    private[crossplatform] protected def myArtifactNameParts: Seq[String] =
+      millModuleSegments.parts.init
+  }
+
+  trait CrossScalaJSModule extends ScalaJSModule with Cross.Module[String] {
+    def crossScalaJSVersion: String = crossValue
+  }
+
+  trait CrossScalaNativeModule
+      extends ScalaNativeModule
+      with Cross.Module[String] {
+    def crossScalaNativeVersion: String = crossValue
+    override def scalaNativeVersion = crossScalaNativeVersion
   }
 }
