+++ mill-crossplatform/src/com/github/lolgab/mill/crossplatform/CrossScalaJSModule.scala
@@ -1,0 +1,8 @@
+package com.github.lolgab.mill.crossplatform
+
+import mill._
+import mill.scalajslib._
+
+trait CrossScalaJSModule extends ScalaJSModule with VersionSpecific.CrossScalaJSModule {
+  override def scalaJSVersion = crossScalaJSVersion
+}
